+++ gems/pending/postgres_ha_admin/postgres_ha_admin.rb
@@ -1,0 +1,13 @@
+require 'util/postgres_dsn_parser'
+
+module PostgresHaAdmin
+
+  class << self
+    attr_writer :logger
+  end
+
+  def self.logger
+    @logger ||= NullLogger.new
+  end
+
+end
